% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Africa_climate}
\alias{Africa_climate}
\title{Climate studies example dataset}
\format{
A dataset with 960 observations and 9 variables.
}
\source{
DOI: 10.24381/cds.adbb2d47 (Accessed on 11-02-2025)
}
\description{
Data extracted from ERA5 hourly data on single levels from 1940 to present
}
\details{
\describe{
\item{Year}{8 years from 1950 to 2020}
\item{Month}{12 calendar months}
\item{Region}{10 IPCC climate reference regions}
\item{AccPrec}{Accumulated precipitation}
\item{DailyEva}{Daily evaporation}
\item{Temp}{Mean temperature}
\item{SoilMois}{Soil moisture}
\item{SPI6}{6-month standardised precipitation index}
\item{wind}{Windspeed}
}
}
\keyword{datasets}
