% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{axes_moveEZ}
\alias{axes_moveEZ}
\title{Provide axes coordinates}
\usage{
axes_moveEZ(bp, which.var)
}
\arguments{
\item{bp}{Object}

\item{which.var}{which variable(s) to find coordinates}
}
\value{
Axes coordinates
}
\description{
Provide axes coordinates
}
