% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{compMeas}
\alias{compMeas}
\title{Measures of comparison based on Orthogonal Procrustes Analysis}
\usage{
compMeas(bp, target = NULL, testee = NULL, centring = TRUE)
}
\arguments{
\item{target}{target configuration}

\item{testee}{testee configuration}

\item{centring}{logical argument to apply centring or not (default is \code{TRUE})}
}
\value{
\item{PS}{Procrustes Statistic}
\item{CC}{Congruence coefficient}
\item{AMB}{Absolute Mean Bias}
\item{MB}{Mean Bias}
\item{RMSB}{Root Mean Squared Bias}
\item{Res.SS}{Residual sum of squares}
\item{Tot.SS}{Total sum of squares}
\item{Fit.SS}{Fitted sum of squares}
}
\description{
Measures of comparison based on Orthogonal Procrustes Analysis
}
