% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.calibrate.axis}
\alias{.calibrate.axis}
\title{Calibrate axis}
\usage{
.calibrate.axis(
  j,
  Xhat,
  means,
  sd,
  axes.rows,
  ax.which,
  ax.tickvec,
  ax.orthogxvec,
  ax.orthogyvec
)
}
\arguments{
\item{j}{j}

\item{Xhat}{Xhat}

\item{means}{means}

\item{sd}{sd}

\item{axes.rows}{axes.rows}

\item{ax.which}{ax.which}

\item{ax.tickvec}{ax.tickvec}

\item{ax.orthogxvec}{ax.orthogxvec}

\item{ax.orthogyvec}{ax.orothogyvec}
}
\value{
Calibrated axes
}
\description{
Calibrate axis
}
