% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{List of evaluation measures}
\usage{
evaluation(bp)
}
\arguments{
\item{bp}{}
}
\value{
list of evaluation measures calculated by \code{compMeas}
}
\description{
List of evaluation measures
}
\examples{
data(Africa_climate)
data(Africa_climate_target)
bp <- biplotEZ::biplot(Africa_climate, scaled = TRUE) |> biplotEZ::PCA()
bp |> moveplot3(time.var = "Year", group.var = "Region", hulls = TRUE,
move = FALSE, target = NULL) |> evaluation()
}
