% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVA testing.R, R/plot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{biplot}
\alias{PCA}
\alias{samples}
\alias{axes}
\alias{fit.measures}
\alias{CVA}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{biplotEZ}{\code{\link[biplotEZ:axes]{axes()}}, \code{\link[biplotEZ:axes]{axes()}}, \code{\link[biplotEZ:biplot]{biplot()}}, \code{\link[biplotEZ:biplot]{biplot()}}, \code{\link[biplotEZ:CVA]{CVA()}}, \code{\link[biplotEZ:fit.measures]{fit.measures()}}, \code{\link[biplotEZ:PCA]{PCA()}}, \code{\link[biplotEZ:PCA]{PCA()}}, \code{\link[biplotEZ:samples]{samples()}}, \code{\link[biplotEZ:samples]{samples()}}}
}}

