% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reflect_biplot}
\alias{reflect_biplot}
\title{Reflect the biplot about a chosen axis}
\usage{
reflect_biplot(bp, reflect.axis = c("FALSE", "x", "y", "xy"))
}
\arguments{
\item{bp}{an object of class \code{biplot}}

\item{reflect.axis}{a character string indicating which axis about to reflect. One of \code{FALSE} (default), "\code{x}" for reflection about the x-axis, "\code{y}" for reflection about the y-axis and "\code{xy}" for reflection about both axes.}
}
\value{
An object of class \code{biplot}
}
\description{
This function provides the user with an option to reflect the biplot horizontally, vertically or diagonally.
}
