% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Measures of comparison for move plot 3}
\usage{
evaluation(bp, centring = TRUE)
}
\arguments{
\item{bp}{biplot object from moveEZ}

\item{centring}{logical argument to apply centring or not (default is \code{TRUE})}
}
\value{
\item{PS}{Procrustes Statistic}
\item{CC}{Congruence coefficient}
\item{AMB}{Absolute Mean Bias}
\item{MB}{Mean Bias}
\item{RMSB}{Root Mean Squared Bias}
\item{Res.SS}{Residual sum of squares}
\item{Tot.SS}{Total sum of squares}
\item{Fit.SS}{Fitted sum of squares}
}
\description{
This function calculates measures of comparison after generalised orthogonal Procrustes Analysis is performed in \code{moveplot3}. Orthogonal Procrustes Analysis is used to compare a target to a testee configuration.
}
\examples{
data(Africa_climate)
data(Africa_climate_target)
bp <- biplotEZ::biplot(Africa_climate, scaled = TRUE) |> biplotEZ::PCA()
bp <- bp |> moveplot3(time.var = "Year", group.var = "Region", hulls = TRUE,
move = FALSE, target = NULL) |> evaluation()

}
