% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveEZ-package.R
\docType{package}
\name{moveEZ-package}
\alias{moveEZ}
\alias{moveEZ-package}
\title{moveEZ: Animated Biplots}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create animated biplots that enables dynamic visualisation of temporal or sequential changes in multivariate data by animating a single biplot across the levels of a time variable. It builds on objects from the 'biplotEZ' package, Lubbe S, le Roux N, Nienkemper-Swanepoel J, Ganey R, Buys R, Adams Z, Manefeldt P (2024) \doi{10.32614/CRAN.package.biplotEZ}, allowing users to create animated biplots that reveal how both samples and variables evolve over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://muvisu.github.io/moveEZ/}
}

}
\author{
\strong{Maintainer}: Raeesa Ganey \email{raeesa.ganey@wits.ac.za} (\href{https://orcid.org/0009-0008-6973-0999}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Johané Nienkemper-Swanepoel \email{nienkemperj@sun.ac.za} (\href{https://orcid.org/0000-0001-6086-8272}{ORCID}) [copyright holder]
}

}
\keyword{internal}
