% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{moveplot}
\alias{moveplot}
\title{Move plot}
\usage{
moveplot(
  bp,
  time.var,
  group.var,
  move = TRUE,
  hulls = TRUE,
  scale.var = 5,
  shadow = FALSE,
  which = NULL,
  label.vars = NULL
)
}
\arguments{
\item{bp}{biplot object from biplotEZ}

\item{time.var}{time variable}

\item{group.var}{group variable}

\item{move}{whether to animate (TRUE) or facet (FALSE) samples, according to time.var}

\item{hulls}{whether to display sample points or convex hulls. A hull requires at least three
non-collinear observations per group per level of \code{time.var}; groups with fewer are
displayed as points instead.}

\item{scale.var}{scaling the vectors representing the variables}

\item{shadow}{whether the animation will keep past states (only when hulls = FALSE)}

\item{which}{integer index into the levels of group.var selecting which groups to display. Default (NULL) shows all groups.}

\item{label.vars}{character vector of column name(s) used to label sample points, pasted together when more than one is supplied. Default (NULL) omits point labels.}
}
\value{
\item{bp}{Returns the elements of the biplot object \code{bp} from \code{biplotEZ}.}
\item{plot}{An animated or a facet of biplots based on the dynamic frame.}
}
\description{
Create animated biplot on samples in a biplot
}
\examples{
data(Africa_climate)
bp <- biplot(Africa_climate, scaled = TRUE) |> PCA()

# Convex hulls facet plot
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = TRUE, move = FALSE)

# Samples facet plot
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = FALSE, move = FALSE)

# Specifying colours with colour palette in biplotEZ
\donttest{
if(interactive()) {
bp <- biplot(Africa_climate, scaled = TRUE, group.aes = Africa_climate$Region) |>
PCA() |> samples(col = RColorBrewer::brewer.pal(10, "Paired"))
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = TRUE, move = FALSE)}}

# Specifying plotting characters for grouping variable in biplotEZ
\donttest{
if(interactive()) {
bp <- biplot(Africa_climate, scaled = TRUE, group.aes = Africa_climate$Region) |>
PCA() |> samples(pch = c(19, 21, 3))
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = FALSE, move = FALSE)}}

# Specifying opacity of plotting characters and size of variable lables
\donttest{
if(interactive()) {
bp <- biplot(Africa_climate, scaled = TRUE, group.aes = Africa_climate$Region) |>
PCA() |> samples(opacity = 0.4) |> axes(label.cex = 1.2)
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = FALSE, move = FALSE)}}

# Specifying colours manually in biplotEZ
\donttest{
if(interactive()) {
bp <- biplot(Africa_climate, scaled = TRUE, group.aes = Africa_climate$Region) |>
PCA() |> samples(col = c("firebrick4", "indianred3", "tomato", "sandybrown",
 "khaki1", "palegreen1", "darkseagreen2", "mediumaquamarine", "deepskyblue4", "mediumpurple4"))
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = TRUE, move = FALSE)}}

# Extracting measures of fit of PCA
bp <- biplot(Africa_climate, scaled = TRUE) |> PCA() |> fit.measures()
bp <- bp |> moveplot(time.var = "Year", group.var = "Region", hulls = TRUE, move = FALSE)
bp$quality
bp$axis.predictivity

# Convex hulls move plot
\donttest{
if(interactive()) {
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = TRUE, move = TRUE)}}

# Samples move plot with shadows
\donttest{
if(interactive()) {
bp |> moveplot(time.var = "Year", group.var = "Region", hulls = FALSE, move = TRUE, shadow = TRUE)}}

# CVA biplot
bp <- biplot(Africa_climate, scaled = TRUE) |> CVA(class = Africa_climate$Region)
bp <- bp |> moveplot2(group.var = "Region", time.var = "Year", move = FALSE)

# Extracting measures of fit of CVA
bp <- biplot(Africa_climate, scaled = TRUE) |>
  CVA(classes = Africa_climate$Region) |> fit.measures()
bp <- bp |> moveplot(time.var = "Year", group.var = "Region", hulls = TRUE, move = FALSE)
bp$quality
bp$axis.predictivity
bp$within.class.axis.predictivity
bp$within.class.sample.predictivity
}
