% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{moveplot}
\alias{moveplot}
\title{Move plot}
\usage{
moveplot(bp, time.var, group.var, moveS = TRUE, hulls = TRUE, scale.var = 5)
}
\arguments{
\item{bp}{biplot object from biplotEZ}

\item{time.var}{time variable}

\item{group.var}{group variable}

\item{moveS}{whether to animate (TRUE) samples or facet (FALSE) samples, according to time.var}

\item{scale.var}{scaling the vectors representing the variables}

\item{hull}{whether to display sample points or convex hulls}
}
\description{
Create animated biplot on samples in a biplot
}
