% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{moveplot}
\alias{moveplot}
\title{Move plot}
\usage{
moveplot(bp, time.var, group.var, move = TRUE, hulls = TRUE, scale.var = 5)
}
\arguments{
\item{bp}{biplot object from biplotEZ}

\item{time.var}{time variable}

\item{group.var}{group variable}

\item{move}{whether to animate (TRUE) or facet (FALSE) samples, according to time.var}

\item{hulls}{whether to display sample points or convex hulls}

\item{scale.var}{scaling the vectors representing the variables}
}
\description{
Create animated biplot on samples in a biplot
}
