% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{moveplot2}
\alias{moveplot2}
\title{Move plot 2}
\usage{
moveplot2(
  bp,
  time.var,
  group.var,
  move = TRUE,
  hulls = TRUE,
  scale.var = 5,
  align.time = NA,
  reflect = NA
)
}
\arguments{
\item{bp}{biplot object from biplotEZ}

\item{time.var}{time variable}

\item{group.var}{group variable}

\item{move}{whether to animate (TRUE) or facet (FALSE) samples and variables, according to time.var}

\item{hulls}{whether to display sample points or convex hulls. A hull requires at least three
non-collinear observations per group per level of \code{time.var}; groups with fewer are
displayed as points instead.}

\item{scale.var}{scaling the vectors representing the variables}

\item{align.time}{a vector specifying the levels of time.var for which the biplots should be aligned. Only biplots corresponding to these time points will be used to compute the alignment transformation.}

\item{reflect}{a character vector specifying the axis of reflection to apply at each corresponding time point in align.time. One of FALSE (default), "x" for reflection about the x-axis, "y" for reflection about the y-axis and "xy" for reflection about both axes.}
}
\value{
\item{bp}{Returns the elements of the biplot object \code{bp} from \code{biplotEZ}.}
\item{plot}{An animated or a facet of biplots based on the dynamic frame.}
}
\description{
Create animated biplot on samples and variables in a biplot
}
\examples{
data(Africa_climate)
bp <- biplot(Africa_climate, scaled = TRUE) |> PCA()
\donttest{
if(interactive()) {
bp <- bp |> moveplot2(time.var = "Year", group.var = "Region", hulls = TRUE, move = TRUE)}}

# Extracting measures of fit for PCA
bp <- bp |> moveplot2(time.var = "Year", group.var = "Region", hulls = TRUE, move = FALSE)
bp$quality
bp$axis.predictivity

# Extracting measures of fit for CVA
bp <- biplot(Africa_climate) |> CVA(classes = Africa_climate$Region)
bp <- bp |> moveplot2(time.var = "Year", group.var = "Region", hulls = TRUE, move = FALSE)
bp$quality
bp$axis.predictivity
bp$within.class.axis.predictivity
bp$within.class.sample.predictivity
}
