% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{moveplot2}
\alias{moveplot2}
\title{Move plot2}
\usage{
moveplot2(bp, time.var, group.var, move = TRUE, hulls = TRUE, scale.var = 5)
}
\arguments{
\item{bp}{biplot object from biplotEZ}

\item{time.var}{time variable}

\item{group.var}{group variable}

\item{move}{whether to animate (TRUE) or facet (FALSE) samples and variables, according to time.var}

\item{scale.var}{scaling the vectors representing the variables}

\item{hull}{whether to display sample points or convex hulls}
}
\description{
Create animated biplot on samples and variables in a biplot
}
